pyproject.toml
[tool.poetry]
name = "play8-backend"
version = "0.1.0"
description = "Play8 Court Machine Booking API"
authors = ["Play8 <info@play8.ai>"]
packages = [{include = "src"}]
package-mode = true

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "0.115.6"
uvicorn = {extras = ["standard"], version = "0.34.0"}
pydantic = "2.10.4"
sqlalchemy = "2.0.36"
psycopg2-binary = "2.9.10"
alembic = "1.14.0"
python-dotenv = "1.0.1"
python-jose = {extras = ["cryptography"], version = "3.3.0"}
httpx = "0.27.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.1.15"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
]

ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__.py

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

